tests/wpt/tests/css/css-values/cap-invalidation.html

<!DOCTYPE html>
<title>CSS Values and Units Test: cap invalidation</title>
<link rel="author" title="Daniil Sakhapov" href="sakhapov@google.com">
<link rel="help" href="https://drafts.csswg.org/css-values/#font-relative-lengths">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  @import url("/fonts/ahem.css");
  html {
    font-family: 'Ahem';
    font-size: 40px;
  }
  div {
    width: 10cap;
  }
</style>

<html>
  <div id="div"></div>
</html>

<script>
  setup({ single_test: true });
  let old_width = div.getBoundingClientRect().width;
  document.documentElement.style.fontFamily = "sans-serif";
  let new_width = div.getBoundingClientRect().width;
  assert_not_equals(old_width, new_width, "expect the capital height of Ahem and sans-serif to be different");
  done();
</script>
